@@ -1,18 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net6.0-windows</TargetFramework>
-    <NoWarn>$(NoWarn);CS1591;IDE0008;CA1812</NoWarn>
-    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-    <NeutralLanguage>en-US</NeutralLanguage>
-    <LangVersion>8.0</LangVersion>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net6.0-windows</TargetFramework>
+		<Nullable>enable</Nullable>
+		<LangVersion>10</LangVersion>
+		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
+		<NoWarn>$(NoWarn);CS1591;IDE0008;CA1812</NoWarn>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.VisualStudio.Extensibility.Sdk" Version="17.4.11-alpha" />
-    <PackageReference Include="Microsoft.VisualStudio.Extensibility.Build" Version="17.4.11-alpha" />
-  </ItemGroup>
-  
-  <ItemGroup>
+		<!-- The VisualStudio.Extensibility preview packages are available from the azure-public/vside/vssdk feed -->
+		<RestoreAdditionalProjectSources>https://pkgs.dev.azure.com/azure-public/vside/_packaging/vssdk/nuget/v3/index.json;$(RestoreAdditionalProjectSources)</RestoreAdditionalProjectSources>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.VisualStudio.Extensibility.Sdk" Version="17.4.15-preview-1" />
+		<PackageReference Include="Microsoft.VisualStudio.Extensibility.Build" Version="17.4.15-preview-1" />
+	</ItemGroup>
+	
+	<ItemGroup>
 		<AdditionalFiles Include=".vsextension\extension.json" ExtensionMetadata="true" />
 	</ItemGroup>
 
@@ -34,10 +37,6 @@
 	<ItemGroup>
 		<Content Include=".vsextension\**\string-resources.json">
 			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-			<IncludeInVsix>true</IncludeInVsix>
 		</Content>
-		<None Update="source.extension.vsixmanifest">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</None>
 	</ItemGroup>
 </Project>
